Update onDrop function and prevent errors in event processing
The commit modifies the onDrop function to simplify its parameters and add flexibility. Meanwhile, to prevent potential errors, null checks for event ids were implemented in the onEventClick and onEventDrop functions of the InteractsWithEvents class. Now, functions will return early if an event id is not set, avoiding potential mishandling of such events.